Change API to separate key and value in src.
Addin an jw_key function and the possibility to separate key and value
insertion to objects lets you use same "value" function for objects and
arrays. One option was considered was to always have an key argument but
ignore it for array insertions, but te seperate key function is more
clear and you can always wrap multiple calls in one function.